requirement.txt
python-docx
win32com
os
tempfile
pythoncom
typing
fastapi